Rename `rad` to `gossip` and explode into smaller modules
