+++ librad/src/net.rs
@@ -1,5 +1,6 @@
 pub mod connection;
 pub mod discovery;
+pub mod gossip;
 pub mod protocol;
 pub mod quic;
 pub mod tls;
